requirements.txt
flask
mysql-connector-python
cpf-utils
